Add Fitter file; separate methods.
